+++ src/Pwa/Response/CmsPageRouteResponse.php
@@ -2,7 +2,6 @@
 
 namespace SwagShopwarePwa\Pwa\Response;
 
-use Shopware\Core\Framework\Struct\Struct;
 use Shopware\Core\System\SalesChannel\StoreApiResponse;
 use SwagShopwarePwa\Pwa\PageResult\AbstractPageResult;